+++ index.js
@@ -1,7 +1,7 @@
 import { useSyncExternalStore } from 'react';
 
 const StateSubFunctions = {
-	__newState(stateName, value){ // defining accessors from initialState
+	__newState(stateName, value){
 		this.__states[stateName] = value;
 		Object.defineProperty(this, stateName, {
 			get(){
@@ -9,37 +9,35 @@ const StateSubFunctions = {
 			},
 			set(value){
 				this.__states[stateName] = value;
-				this.__broadcast(stateName); // broadcast to subscribers after any setting a new value to state. 
+				this.__broadcast(stateName);
 			}
 		});
 	},
-
 	__subscribe(stateName, f){
-		if(!(stateName in this.__subsribers)){ // create a new room for subscribers, if it doesnt exist
+		if(!(stateName in this.__subsribers)){
 			this.__subsribers[stateName] = new Set();
 		}
-		this.__subsribers[stateName].add(f); // adding a subscriber to a room
-		return ()=> this.__subsribers[stateName].delete(f); // return "subCleaner" for useSyncExternalStore (used when unmounting the component)
+		this.__subsribers[stateName].add(f);
+		return ()=> this.__subsribers[stateName].delete(f);
 	},
-
 	__broadcast(stateName){
 		if(stateName in this.__subsribers){
 			this.__subsribers[stateName].forEach(f => f());
 		}
 	},
-
 	__useSyncExternalStore(stateName){
 		return useSyncExternalStore(this.__subscribe.bind(this, stateName), () => this.__states[stateName]);
 	},
-
 	__states: {},
-
 	__subsribers: {},
-
+	subTo(stateName){ // will be deprecated in the future
+		return stateName in this.__states ? this.__useSyncExternalStore(stateName) : undefined;
+	}
 };
 
-class StateSub{
+class StateSub extends Function{
 	constructor(store){
+		super();
 		StateSubFunctions.__proto__ = new Proxy(store, {
 			get(target, key, receiver){
 				let value = target[key];
@@ -48,15 +46,12 @@ class StateSub{
 		});
 		this.__proto__ = StateSubFunctions;
 		for(let state in store.initialState){
-			this.__newState(state, store.initialState[state]); // transferring initial states to store
+			this.__newState(state, store.initialState[state]);
 		}
-		delete store.initialState; // clearing initial states
-		Object.defineProperty(this, 'subTo', { // defining the subscription function
-			value: function(stateName){
-				if(stateName in this.__states){
-					return this.__useSyncExternalStore(stateName);
-				}
-				return undefined; // when someone tries to subscribe to an undefined state, the subscription will be skipped and returned "undefined"
+		delete store.initialState;
+		return new Proxy(this, {
+			apply: (target, thisArg, stateName) => {
+				return stateName in target.__states ? target.__useSyncExternalStore(stateName) : undefined;
 			}
 		});
 	}