+++ CHANGELOG.md
@@ -1,4 +1,21 @@
-# 0.0.5 (2022-09-05)
+# Changelog
+
+## 0.1.0
+
+Now it's easier to subscribe to a state change.
+#### Before
+```jsx
+const list = store.subTo('list');
+```
+
+#### Now
+```jsx
+const list = store('list');
+```
+
+Method "subTo" will be deprecated in the future.
+
+## 0.0.5 (2022-09-05)
 
 Added context auto-binding to store functions.<br/>
 For example, now you don't need to bind the context in handlers.
@@ -27,10 +44,10 @@ const Button = () => {
 ```
 
 
-# 0.0.4 (2022-09-01)
+## 0.0.4 (2022-09-01)
 
-Added demos<br/>
-Bug fixed<br/>
+Added demos.<br/>
+Bug fixed.<br/>
 Removed a private object properties.<br/>
 
 #### Before